fix: correçoes de conflitos gerados por commits

+++ src/main/java/com/anatonelly/freteexpress/service/FreteService.java
@@ -3,7 +3,7 @@ package com.anatonelly.freteexpress.service;
 import com.anatonelly.freteexpress.model.EmpresaCliente;
 import com.anatonelly.freteexpress.model.Frete;
 import com.anatonelly.freteexpress.model.Motorista;
-import com.anatonelly.freteexpress.enums.StatusFrete; // Certifique-se que este enum existe e tem os valores
+import com.anatonelly.freteexpress.enums.StatusFrete;
 import com.anatonelly.freteexpress.repository.FreteRepository;
 import com.anatonelly.freteexpress.repository.MotoristaRepository;
 import org.springframework.stereotype.Service;
@@ -26,9 +26,8 @@ public class FreteService {
         return freteRepository.findAll();
     }
 
-    // O ID da entidade Frete deve ser Long, se você usou Long no JpaRepository
-    // Se o ID for Integer, ajuste aqui também
-    public Optional<Frete> buscarPorId(Integer id) { // Alterado para Integer
+    // Corrigido para usar Long, correspondendo ao tipo do ID da entidade Frete
+    public Optional<Frete> buscarPorId(Long id) {
         return freteRepository.findById(id);
     }
 
@@ -36,7 +35,8 @@ public class FreteService {
         return freteRepository.save(frete);
     }
 
-    public boolean deletar(Integer id) { // Alterado para Integer
+    // Corrigido para usar Long
+    public boolean deletar(Long id) {
         if (freteRepository.existsById(id)) {
             freteRepository.deleteById(id);
             return true;
@@ -44,22 +44,21 @@ public class FreteService {
         return false;
     }
 
-    // O ID da entidade Frete deve ser Long, se você usou Long no JpaRepository
-    // Se o ID for Integer, ajuste aqui também
-    public Optional<Frete> atualizar(Integer id, Frete freteAtualizado) { // Alterado para Integer
+    // Corrigido para usar Long
+    public Optional<Frete> atualizar(Long id, Frete freteAtualizado) {
         return freteRepository.findById(id).map(freteExistente -> {
-            // Certifique-se que os getters/setters abaixo estão na sua classe Frete.java
-            freteExistente.setOrigem(freteAtualizado.getOrigem()); // Exemplo, verificar se existe na classe Frete
-            freteExistente.setDestino(freteAtualizado.getDestino()); // Exemplo
-            freteExistente.setTipoCarga(freteAtualizado.getTipoCarga()); // Exemplo
-            freteExistente.setPeso(freteAtualizado.getPeso()); // Exemplo
-            freteExistente.setDimensoes(freteAtualizado.getDimensoes()); // Exemplo
-            freteExistente.setPrazoEntrega(freteAtualizado.getPrazoEntrega()); // Exemplo
-            freteExistente.setDescricaoCarga(freteAtualizado.getDescricaoCarga()); // Exemplo
-            freteExistente.setValorFrete(freteAtualizado.getValorFrete()); // Exemplo
-            freteExistente.setStatus(freteAtualizado.getStatus()); // Exemplo
-            freteExistente.setEmpresaCliente(freteAtualizado.getEmpresaCliente()); // Exemplo
-            freteExistente.setMotoristaSolicitante(freteAtualizado.getMotoristaSolicitante()); // Exemplo
+            // A lógica de atualização dos campos permanece aqui
+            freteExistente.setOrigem(freteAtualizado.getOrigem());
+            freteExistente.setDestino(freteAtualizado.getDestino());
+            freteExistente.setTipoCarga(freteAtualizado.getTipoCarga());
+            freteExistente.setPeso(freteAtualizado.getPeso());
+            freteExistente.setDimensoes(freteAtualizado.getDimensoes());
+            freteExistente.setPrazoEntrega(freteAtualizado.getPrazoEntrega());
+            freteExistente.setDescricaoCarga(freteAtualizado.getDescricaoCarga());
+            freteExistente.setValorFrete(freteAtualizado.getValorFrete());
+            freteExistente.setStatus(freteAtualizado.getStatus());
+            freteExistente.setEmpresaCliente(freteAtualizado.getEmpresaCliente());
+            freteExistente.setMotoristaSolicitante(freteAtualizado.getMotoristaSolicitante());
             return freteRepository.save(freteExistente);
         });
     }
@@ -72,15 +71,17 @@ public class FreteService {
         return freteRepository.findByEmpresaClienteOrderByPrazoEntregaDesc(empresaCliente);
     }
 
-    public boolean solicitarFrete(Integer freteId, Integer motoristaId) { // <<<<< CORRIGIDO AQUI: Long para Integer
+    // Corrigido para usar Long para ambos os IDs
+    // CORREÇÃO FINAL: O ID do frete é Long, e o ID do motorista é Integer.
+    public boolean solicitarFrete(Long freteId, Integer motoristaId) {
         Optional<Frete> freteOpt = freteRepository.findById(freteId);
         Optional<Motorista> motoristaOpt = motoristaRepository.findById(motoristaId);
 
         if (freteOpt.isPresent() && motoristaOpt.isPresent()) {
             Frete frete = freteOpt.get();
-            if (frete.getStatus() == StatusFrete.PENDENTE) { // Certifique-se que StatusFrete.PENDENTE existe
-                frete.setStatus(StatusFrete.SOLICITADO); // Certifique-se que StatusFrete.SOLICITADO existe
-                frete.setMotoristaSolicitante(motoristaOpt.get()); // Certifique-se que este setter existe em Frete
+            if (frete.getStatus() == StatusFrete.PENDENTE) {
+                frete.setStatus(StatusFrete.SOLICITADO);
+                frete.setMotoristaSolicitante(motoristaOpt.get());
                 freteRepository.save(frete);
                 return true;
             }
